+++ src/module/preloadTemplates.ts
@@ -10,6 +10,8 @@ export async function preloadTemplates() {
     "modules/midi-qol/templates/roll-stats.html",
     "modules/midi-qol/templates/lateTargeting.html",
     // "modules/midi-qol/templates/midiProperties.html"
+    "modules/midi-qol/templates/sound-config.html",
+    "modules/midi-qol/templates/rollAlternate.html",
   ];
 	return loadTemplates(templatePaths);
 }